routes: add handling of root servers
Root Servers don't have routes in that sense. But to let them join the
cluster successfully we need to make k8s think it has proper routes.

This is done by propagating the routes based on the root server's pod
CIDR and internal IP.

This is by no means a full solution. It just works for our internal use
case.